+++ CHANGELOG.md
@@ -17,6 +17,11 @@
 
 ### Fixed
 
+- Send protocol-level pings on idle browser WebSockets and close connections that do not return peer
+  traffic within 15 seconds, removing dead transports without claiming to detect frozen JavaScript.
+  [PR #93](https://github.com/kcosr/herdr-web/pull/93), contributed by
+  [Will Hampson (@Whamp)](https://github.com/Whamp).
+
 ### Removed
 
 ## [0.6.1] - 2026-09-12